Create workflow for XRechnung Validator
I wanted to validate XML documents in the UBL standard,
or more specifically XRechnung.
I ran the validator locally. [1]
The validator configuration was based on XML Schema
and Schematron. [2]
I managed to put it all together thanks to the documentation [3]
I found.

Then I analyzed which elements were actually needed
and built a GitHub Actions workflow.
This allows me to run the validator for any XML file in the repository.
When validation fails, I upload error reports as artifacts.

[1]: https://github.com/itplr-kosit/validator
[2]: https://github.com/itplr-kosit/validator-configuration-xrechnung
[3]: https://github.com/itplr-kosit/validator-configuration-xrechnung/blob/master/docs/usage.md